requirements.txt
python-dotenv==1.0.1
tweepy==4.15.0
praw==7.8.1